db 함수를 레거시로 변경함.
