+++ db_item.go
@@ -1647,37 +1647,6 @@ func Type(session *mgo.Session, project, name string) (string, error) {
 	return items[0].Type, nil
 }
 
-// SetImageSize 함수는 해당 샷의 이미지 사이즈를 설정한다. // legacy
-// key 설정값 : platesize, undistortionsize, rendersize
-func SetImageSize(session *mgo.Session, project, name, key, size string) (string, error) {
-	if !(key == "platesize" || key == "dsize" || key == "undistortionsize" || key == "rendersize") {
-		return "", errors.New("잘못된 key값입니다")
-	}
-	session.SetMode(mgo.Monotonic, true)
-	err := HasProject(session, project)
-	if err != nil {
-		return "", err
-	}
-	typ, err := Type(session, project, name)
-	if err != nil {
-		return "", err
-	}
-	id := name + "_" + typ
-	c := session.DB("project").C(project)
-	if key == "dsize" || key == "undistortionsize" {
-		err = c.Update(bson.M{"id": id}, bson.M{"$set": bson.M{"dsize": size, "undistortionsize": size, "updatetime": time.Now().Format(time.RFC3339)}})
-		if err != nil {
-			return id, err
-		}
-	} else {
-		err = c.Update(bson.M{"id": id}, bson.M{"$set": bson.M{key: size, "updatetime": time.Now().Format(time.RFC3339)}})
-		if err != nil {
-			return id, err
-		}
-	}
-	return id, nil
-}
-
 // SetImageSizeVer2 함수는 해당 샷의 이미지 사이즈를 설정한다.
 // key 설정값 : platesize, undistortionsize, rendersize
 func SetImageSizeVer2(session *mgo.Session, project, id, key, size string) error {
